requirements.txt
ExifRead==2.1.2
setuptools==18.2
Pillow==3.3.1
